[Bronze V] Title: 이상한 기호, Time: 116 ms, Memory: 21084 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.kt
@@ -1,5 +1,5 @@
 fun main() {
-    val (A, B) = readln().split(" ").map { it.toInt() }
+    val (A, B) = readln().split(" ").map { it.toLong() }
     val result = (A + B) * (A - B)
     println(result)
 }
